requirements.txt
httpx>=0.27.0
aiohttp>=3.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0
